requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiorate==1.0.2
aiosignal==1.3.1
anyio==4.6.2.post1
aptos-sdk==0.10.0
attrs==24.2.0
base58==2.1.1
behave==1.2.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
construct==2.10.68
construct-typing==0.5.6
coverage==7.6.10
ecdsa==0.18.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonalias==0.1.1
multidict==6.1.0
nest-asyncio==1.6.0
parse==1.20.2
parse_type==0.6.4
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pycryptodome==3.21.0
PyNaCl==1.5.0
pyTelegramBotAPI==4.24.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-graphql-client
pytz==2024.2
requests==2.32.3
segno==1.6.1
six==1.16.0
sniffio==1.3.1
telebot==0.0.5
tomli==2.1.0
tornado==6.4.2
types-six==1.16.21.20241105
typing_extensions==4.12.2
tweepy==4.15.0
tzdata==2025.1
urllib3==2.2.3
websockets
yarl==1.17.2
base58==2.1.1
cryptography==44.0.0
numpy==2.2.1
pandas==2.2.3
psycopg2==2.9.10
python-telegram-bot==21.10
Requests==2.32.3
solana==0.36.2
solders==0.23.0
python-telegram-bot[webhooks]
python-dotenv
